+++ engine/core/windows/SystemWin.cpp
@@ -40,11 +40,11 @@ namespace ouzel::core::windows
                 {
                     const auto bufferSize = WideCharToMultiByte(CP_UTF8, 0, argv[i], -1, nullptr, 0, nullptr, nullptr);
                     if (bufferSize == 0)
-                        throw std::system_error{GetLastError(), std::system_category(), "Failed to convert wide char to UTF-8"};
+                        throw std::system_error{static_cast<int>(GetLastError()), std::system_category(), "Failed to convert wide char to UTF-8"};
 
                     auto buffer = std::make_unique<char[]>(bufferSize);
                     if (WideCharToMultiByte(CP_UTF8, 0, argv[i], -1, buffer.get(), bufferSize, nullptr, nullptr) == 0)
-                        throw std::system_error{GetLastError(), std::system_category(), "Failed to convert wide char to UTF-8"};
+                        throw std::system_error{static_cast<int>(GetLastError()), std::system_category(), "Failed to convert wide char to UTF-8"};
 
                     result.push_back(buffer.get());
                 }
