+++ engine/graphics/opengl/windows/OGLRenderDeviceWin.cpp
@@ -32,7 +32,7 @@ namespace ouzel::graphics::opengl::windows
             {
                 HINSTANCE instance = GetModuleHandleW(nullptr);
                 if (!instance)
-                    throw std::system_error{GetLastError(), std::system_category(), "Failed to get module handle"};
+                    throw std::system_error{static_cast<int>(GetLastError()), std::system_category(), "Failed to get module handle"};
 
                 WNDCLASSW wc;
                 wc.style = CS_OWNDC;
@@ -48,14 +48,14 @@ namespace ouzel::graphics::opengl::windows
 
                 windowClass = RegisterClassW(&wc);
                 if (!windowClass)
-                    throw std::system_error{GetLastError(), std::system_category(), "Failed to register window class"};
+                    throw std::system_error{static_cast<int>(GetLastError()), std::system_category(), "Failed to register window class"};
 
                 window = CreateWindowW(MAKEINTATOM(windowClass), L"TempWindow", 0,
                                        CW_USEDEFAULT, CW_USEDEFAULT,
                                        CW_USEDEFAULT, CW_USEDEFAULT,
                                        0, 0, instance, 0);
                 if (!window)
-                    throw std::system_error{GetLastError(), std::system_category(), "Failed to create window"};
+                    throw std::system_error{static_cast<int>(GetLastError()), std::system_category(), "Failed to create window"};
 
                 deviceContext = GetDC(window);
                 if (!deviceContext)
@@ -91,17 +91,17 @@ namespace ouzel::graphics::opengl::windows
 
                 const auto pixelFormat = ChoosePixelFormat(deviceContext, &pixelFormatDesc);
                 if (!pixelFormat)
-                    throw std::system_error{GetLastError(), std::system_category(), "Failed to choose pixel format"};
+                    throw std::system_error{static_cast<int>(GetLastError()), std::system_category(), "Failed to choose pixel format"};
 
                 if (!SetPixelFormat(deviceContext, pixelFormat, &pixelFormatDesc))
-                    throw std::system_error{GetLastError(), std::system_category(), "Failed to set pixel format"};
+                    throw std::system_error{static_cast<int>(GetLastError()), std::system_category(), "Failed to set pixel format"};
 
                 renderContext = wglCreateContext(deviceContext);
                 if (!renderContext)
-                    throw std::system_error{GetLastError(), std::system_category(), "Failed to create OpenGL rendering context"};
+                    throw std::system_error{static_cast<int>(GetLastError()), std::system_category(), "Failed to create OpenGL rendering context"};
 
                 if (!wglMakeCurrent(deviceContext, renderContext))
-                    throw std::system_error{GetLastError(), std::system_category(), "Failed to set OpenGL rendering context"};
+                    throw std::system_error{static_cast<int>(GetLastError()), std::system_category(), "Failed to set OpenGL rendering context"};
             }
 
             ~TempContext()
@@ -253,17 +253,17 @@ namespace ouzel::graphics::opengl::windows
             UINT numFormats;
 
             if (!wglChoosePixelFormatProc(deviceContext, attributeList, nullptr, 1, &pixelFormat, &numFormats))
-                throw std::system_error{GetLastError(), std::system_category(), "Failed to choose pixel format"};
+                throw std::system_error{static_cast<int>(GetLastError()), std::system_category(), "Failed to choose pixel format"};
         }
         else
         {
             pixelFormat = ChoosePixelFormat(deviceContext, &pixelFormatDesc);
             if (!pixelFormat)
-                throw std::system_error{GetLastError(), std::system_category(), "Failed to choose pixel format"};
+                throw std::system_error{static_cast<int>(GetLastError()), std::system_category(), "Failed to choose pixel format"};
         }
 
         if (!SetPixelFormat(deviceContext, pixelFormat, &pixelFormatDesc))
-            throw std::system_error{GetLastError(), std::system_category(), "Failed to set pixel format"};
+            throw std::system_error{static_cast<int>(GetLastError()), std::system_category(), "Failed to set pixel format"};
 
         if (wglCreateContextAttribsProc)
         {
@@ -292,7 +292,7 @@ namespace ouzel::graphics::opengl::windows
             throw std::runtime_error{"Failed to create OpenGL rendering context"};
 
         if (!wglMakeCurrent(deviceContext, renderContext))
-            throw std::system_error{GetLastError(), std::system_category(), "Failed to set OpenGL rendering context"};
+            throw std::system_error{static_cast<int>(GetLastError()), std::system_category(), "Failed to set OpenGL rendering context"};
 
         apiVersion = getVersion();
 
@@ -303,7 +303,7 @@ namespace ouzel::graphics::opengl::windows
              static_cast<GLsizei>(window.getResolution().v[1]));
 
         if (!wglMakeCurrent(deviceContext, nullptr))
-            throw std::system_error{GetLastError(), std::system_category(), "Failed to unset OpenGL rendering context"};
+            throw std::system_error{static_cast<int>(GetLastError()), std::system_category(), "Failed to unset OpenGL rendering context"};
     }
 
     RenderDevice::~RenderDevice()
@@ -338,7 +338,7 @@ namespace ouzel::graphics::opengl::windows
     void RenderDevice::present()
     {
         if (!SwapBuffers(deviceContext))
-            throw std::system_error{GetLastError(), std::system_category(), "Failed to swap buffers"};
+            throw std::system_error{static_cast<int>(GetLastError()), std::system_category(), "Failed to swap buffers"};
     }
 
     void RenderDevice::renderMain()
@@ -346,7 +346,7 @@ namespace ouzel::graphics::opengl::windows
         thread::setCurrentThreadName("Render");
 
         if (!wglMakeCurrent(deviceContext, renderContext))
-            throw std::system_error{GetLastError(), std::system_category(), "Failed to set OpenGL rendering context"};
+            throw std::system_error{static_cast<int>(GetLastError()), std::system_category(), "Failed to set OpenGL rendering context"};
 
         while (running)
         {
@@ -361,7 +361,7 @@ namespace ouzel::graphics::opengl::windows
         }
 
         if (!wglMakeCurrent(deviceContext, nullptr))
-            throw std::system_error{GetLastError(), std::system_category(), "Failed to unset OpenGL rendering context"};
+            throw std::system_error{static_cast<int>(GetLastError()), std::system_category(), "Failed to unset OpenGL rendering context"};
     }
 }
